Move duplicate definition into util header
